Fix edge discovery for "touching nodes"
There was an issue with edge discovery when the nodes were "touching"
because np.nonzero()[0] would return index with value '0'. '0' is
perceived by python as False because of that this would not pass the
if statement that is needed to be passed to add the add the edge
to node_degrees.