Change the init value of node id counter
There was an issue with indexing during training when node indexing
was starting at 0.

This patch fixes this by changing the initial value of node counter to
0. But to be able to do that we need to fix the node-visibility because
we are not able to find node with value 0 in the render_image field.

To fix this we increase the value of id stored to the array and
then we decrease it (not nice but the most effective solution right now).